+++ lib/src/chat_context_menu_wrapper.dart
@@ -13,17 +13,62 @@ class ChatContextMenuWrapper extends StatefulWidget {
     this.padding = const EdgeInsets.all(8),
     this.requestFocus = false,
     this.shadows,
+    this.arrowHeight = 8.0,
+    this.arrowWidth = 12.0,
+    this.spacing = 10.0,
     this.transitionsBuilder,
   });
 
+  ///在页面中显示的组件
+  ///Components displayed on the page
   final ContextMenuWidgetBuilder widgetBuilder;
+
+  ///显示的 context menu 组件
+  ///Displayed context menu widget
   final ContextMenuContentBuilder menuBuilder;
+
+  ///屏障颜色 （背景叠加层的颜色。）
+  ///Color of the background overlay.
   final Color barrierColor;
+
+  ///context menu容器的背景颜色。
+  ///也作用于角标的颜色
+  ///Background color of the menu container.
+  ///Also affects the color of the marker
   final Color? backgroundColor;
+
+  ///context menu容器的圆角
+  ///Rounded corners of context menu container
   final BorderRadius borderRadius;
+
+  ///context menu容器的内边距
+  ///Padding corners of context menu container
   final EdgeInsets padding;
+
+  ///是否在显示的的时候获取焦点
+  ///默认为 false 如果为 true 那么在context menu显示的时候你页面中的其他具有焦点的组件将失去焦点
+  ///Whether to get focus when displayed
+  ///The default is false. If it is true, other focused components in your page will lose focus when the context menu is displayed.
   final bool requestFocus;
+
+  ///context menu容器的阴影
+  ///Shadows corners of context menu container
   final List<BoxShadow>? shadows;
+
+  ///角标的高度
+  ///The height of the marker
+  final double arrowHeight;
+
+  ///角标的宽度
+  ///The width of the marker
+  final double arrowWidth;
+
+  ///context menu 和组件的间距
+  ///Spacing between context menu and components
+  final double spacing;
+
+  ///自定义出现的动画
+  ///Customize the animation that appears
   final Widget? Function(
     BuildContext context,
     Animation<double> animation,
@@ -55,6 +100,9 @@ class _ChatContextMenuWrapperState extends State<ChatContextMenuWrapper> {
       requestFocus: widget.requestFocus,
       padding: widget.padding,
       shadows: widget.shadows,
+      arrowHeight: widget.arrowHeight,
+      arrowWidth: widget.arrowWidth,
+      spacing: widget.spacing,
       transitionsBuilder: widget.transitionsBuilder,
     );